categorize windows.skeleton_key_check as malware plugin
